base: add basic networking support and ircd module
Adds support for allowing subjects to send and receive messages to /
from ports associated with generic network interfaces and network nodes
(i.e., node_t and netif_t, aliased to networking.generic_node and
networking.generic_interface respectively).

ngircd is an example of a simple networked system service which binds to
a private port and listens for incoming connections, so it's a good
candidate for an example module for porting from kernel policy to CIL.
It also already has a private port type defined by the corenetwork base
module, and is a good opportunitiy of building a bridge between
refpolicy and CIL.

The port, node, and interface macros in the network module currently
only support the tcp socket class.  This will be refactored to support
the various other socket classes later.

Also move the example 'my_system_service' module out of the contrib/
tree and into an examples/ tree, now that there is an ircd module to use
as a contrib reference.